+++ CHANGELOG.md
@@ -5,6 +5,9 @@
 - `SchemaGenerator::definitions_mut()` which returns a mutable reference to the generator's schema definitions
 
 ### Changed:
+- Minimum supported rust version is now 1.37.0
+
+### Changed:
 - Deriving JsonSchema on enums now sets `additionalProperties` to false on generated schemas wherever serde doesn't accept unknown properties. This includes non-unit variants of externally tagged enums, and struct-style variants of all enums that have the `deny_unknown_fields` attribute.
 
 ## [0.8.0] - 2020-09-27